api/requirements.txt
sqlalchemy
fastapi
uvicorn[standard]
fastapi[standard]
python-dotenv
passlib[argon2]